+++ molo/polls/models.py
@@ -1,21 +1,47 @@
 from django.db import models
-from wagtail.wagtailcore.models import Page
-from wagtail.wagtailadmin.edit_handlers import (
-    FieldPanel, MultiFieldPanel, FieldRowPanel)
-from molo.core.models import ArticlePage, SectionPage, TranslatablePageMixin
+from django.dispatch import receiver
 from django.utils.translation import ugettext_lazy as _
 
+from wagtail.wagtailcore.models import Page
+from wagtail.wagtailadmin.edit_handlers import (
+    FieldPanel, MultiFieldPanel, FieldRowPanel,
+)
+
+from molo.core.utils import generate_slug
+from molo.core.models import (
+    Main, ArticlePage, SectionPage, TranslatablePageMixinNotRoutable,
+    PreventDeleteMixin, index_pages_after_copy,
+)
 
 SectionPage.subpage_types += ['polls.Question', 'polls.FreeTextQuestion']
 ArticlePage.subpage_types += ['polls.Question', 'polls.FreeTextQuestion']
 
 
-class PollsIndexPage(Page):
-    parent_page_types = []
+class PollsIndexPage(Page, PreventDeleteMixin):
+    parent_page_types = ['core.Main']
     subpage_types = ['polls.Question', 'polls.FreeTextQuestion']
 
+    def copy(self, *args, **kwargs):
+        site = kwargs['to'].get_site()
+        main = site.root_page
+        PollsIndexPage.objects.child_of(main).delete()
+        super(PollsIndexPage, self).copy(*args, **kwargs)
 
-class Question(TranslatablePageMixin, Page):
+
+@receiver(index_pages_after_copy, sender=Main)
+def create_polls_index_page(sender, instance, **kwargs):
+    if not instance.get_children().filter(
+            title='Polls').exists():
+        polls_index = PollsIndexPage(
+            title='Polls', slug=('polls-%s' % (
+                generate_slug(instance.title), )))
+        instance.add_child(instance=polls_index)
+        polls_index.save_revision().publish()
+
+
+class Question(TranslatablePageMixinNotRoutable, Page):
+    parent_page_types = [
+        'polls.PollsIndexPage', 'core.SectionPage', 'core.ArticlePage']
     subpage_types = ['polls.Choice']
     short_name = models.TextField(
         null=True, blank=True,
@@ -83,6 +109,7 @@ class Question(TranslatablePageMixin, Page):
                     parent_article  .get_effective_extra_style_hints()
         return self.extra_style_hints
 
+
 Question.settings_panels = [
     MultiFieldPanel(
         [FieldRowPanel(
@@ -92,6 +119,8 @@ Question.settings_panels = [
 
 
 class FreeTextQuestion(Question):
+    parent_page_types = [
+        'polls.PollsIndexPage', 'core.SectionPage', 'core.ArticlePage']
     subpage_types = []
     content_panels = Page.content_panels
     numerical = models.BooleanField(
@@ -111,7 +140,8 @@ class FreeTextQuestion(Question):
             user=user, question__id=self.get_main_language_page().id).exists())
 
 
-class Choice(TranslatablePageMixin, Page):
+class Choice(TranslatablePageMixinNotRoutable, Page):
+    parent_page_types = ['polls.Question']
     subpage_types = []
     votes = models.IntegerField(default=0)
     choice_votes = models.ManyToManyField('ChoiceVote',