+++ molo/polls/templatetags/poll_votings.py
@@ -15,7 +15,8 @@ register = template.Library()
 def poll_page(context, pk=None):
     context = copy(context)
     locale_code = context.get('locale_code')
-    page = PollsIndexPage.objects.live().first()
+    main = context['request'].site.root_page
+    page = PollsIndexPage.objects.child_of(main).live().first()
     if page:
         questions = (
             Question.objects.child_of(page).filter(
@@ -27,6 +28,23 @@ def poll_page(context, pk=None):
         'questions': get_pages(context, questions, locale_code)
     })
     return context
+
+
+@register.assignment_tag(takes_context=True)
+def load_polls(context):
+    context = copy(context)
+    locale_code = context.get('locale_code')
+    main = context['request'].site.root_page
+    page = PollsIndexPage.objects.child_of(main).live().first()
+
+    if page:
+        questions = (
+            Question.objects.child_of(page).filter(
+                languages__language__is_main_language=True).specific())
+    else:
+        questions = Question.objects.none()
+
+    return get_pages(context, questions, locale_code)
 
 
 @register.inclusion_tag('polls/poll_page_in_section.html',